primeira iteração caminhos e desvios com ProcessTreeVisualisationInfo

+++ src/main/java/projeto/algorithms_process_mining/inductive_miner/IvMHelper.java
@@ -17,7 +17,7 @@ import java.util.List;
 @NoArgsConstructor
 public class IvMHelper {
     private HashMap<Integer,Long> activityCardinalitiesComplete;
-    private HashMap<Pair<Integer,Integer>,Long> edgeCardinality;
+    private HashMap<Pair<Integer,Integer>,Integer> edgeCardinality;
     private IvMModel model;
     private List<Pair<Integer,Integer>> startActivities;
     private List<Pair<Integer,Integer>> endActivities;
